@@ -1,19 +1,19 @@
-import { createRouter, createWebHistory } from "vue-router";
-import Home from "@/views/Home/index.vue";
-import Login from "@/views/Login/index.vue";
+import { createRouter, createWebHistory } from 'vue-router'
+import Home from '@/views/Home/index.vue'
+import Login from '@/views/Login/index.vue'
 
 const routes = [
   {
-    path: "/",
-    redirect: "/home",
+    path: '/',
+    redirect: '/home',
   },
-  { path: "/home", name: "home", component: Home },
-  { path: "/login", name: "login", component: Login },
-];
+  { path: '/home', name: 'home', component: Home },
+  { path: '/login', name: 'login', component: Login },
+]
 
 const router = createRouter({
   history: createWebHistory(),
   routes,
-});
+})
 
-export default router;
+export default router
