@@ -1,8 +1,8 @@
 /// <reference types="vite/client" />
-declare module "*.vue" {
-  import { DefineComponent } from "vue";
-  const component: DefineComponent<{}, {}, any>;
-  export default component;
+declare module '*.vue' {
+  import { DefineComponent } from 'vue'
+  const component: DefineComponent<{}, {}, any>
+  export default component
 }
 
-declare module "element-plus/dist/locale/zh-cn.mjs";
+declare module 'element-plus/dist/locale/zh-cn.mjs'
